# PERF: Improve performance of CategoricalIndex.is_unique