@topper-123 2018-05-19
``is_unique`` is a cached property, so I don't think an ASV makes sense in this case. (I assume ``setup`` is called once for each time method, but not once for every call to the time method, haven't checked)