@topper-123 2018-05-20
It seems to me that ASV does "best of X" style of testing. from the [docs](https://asv.readthedocs.io/en/latest/writing_benchmarks.html#timing):

> The timing itself is based on the Python standard library’s timeit module, with some extensions for automatic heuristics shamelessly stolen from IPython’s %timeit magic function

I could add the timing tests, but I assume these improvement is not seen well in repeated calls to ``is_unique``. You want them added still?