@jreback 2018-05-21
we do have asv's for others. sure if can tag so these are run once is better (not sure if that is possible)


```
(pandas) bash-3.2$ grep -r is_unique asv_bench/benchmarks/
asv_bench/benchmarks//algorithms.py:        # cache is_unique
asv_bench/benchmarks//algorithms.py:        self.idx_int_dup.is_unique
```