@topper-123 2018-06-04
> we do have asv's for others. sure if can tag so these are run once is better (not sure if that is possible)
>
> (pandas) bash-3.2$ grep -r is_unique asv_bench/benchmarks/
> asv_bench/benchmarks//algorithms.py:        # cache is_unique
> asv_bench/benchmarks//algorithms.py:        self.idx_int_dup.is_unique

Those two are in a setup method, so timing are not done on them:

```python
class DuplicatedUniqueIndex(object):

    goal_time = 0.2

    def setup(self):
        N = 10**5
        self.idx_int_dup = pd.Int64Index(np.arange(N * 5))
        # cache is_unique
        self.idx_int_dup.is_unique

     def time_duplicated_unique_int(self):
```

So currently there is no tests for ``is_unique`` in asv, and it probably doesn't make sense  IMO, as ``is_unique`` is a cached property and timing it will just return the cached value, not the algorithm...