@jreback 2018-06-04
no i think would need to create one for purposes of measuring cached properties